🐛

+++ database/migrations/2023_12_14_102608_add_event_column_to_activity_log_table.php
@@ -9,14 +9,18 @@ class AddEventColumnToActivityLogTable extends Migration
     public function up()
     {
         Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->string('event')->nullable()->after('subject_type');
+            if (DB::getDriverName() !== 'mongodb') {
+                $table->string('event')->nullable()->after('subject_type');
+            }
         });
     }
 
     public function down()
     {
         Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->dropColumn('event');
+            if (DB::getDriverName() !== 'mongodb') {
+                $table->dropColumn('event');
+            }
         });
     }
 }
